#3385 modal outline progress
