+++ src/frontend/src/pages/ProjectDetailPage/ProjectViewContainer/PartReview/PartReviewComponents/ReviewFormView.tsx
@@ -1,70 +1,8 @@
-import { Control, Controller, FieldErrors, UseFormHandleSubmit } from 'react-hook-form';
+import { Control, FieldErrors, UseFormHandleSubmit, UseFormSetValue, UseFormReset } from 'react-hook-form';
 import NERFormModal from '../../../../../components/NERFormModal';
 import { FormControl, FormLabel } from '@mui/material';
 import ReactHookTextField from '../../../../../components/ReactHookTextField';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { useForm } from 'react-hook-form';
-import * as yup from 'yup';
-
-const schema = yup.object().shape({
-  partId: yup.string().required('Select a Part!'),
-  fileIds: yup.array().of(yup.string().defined()).required(), //????
-  submissionId:  yup.string().required('Select a Submission!'),
-  notes: yup.string().optional()
-});
-
-export interface ReviewFormInput {
-  partId: string;
-  fileIds: string[];
-  submissionId: string;
-  notes?: string;
-}
-
-export interface ReviewFormProps {
-  submitText: 'Add' | 'Edit';
-  onSubmit: (payload: ReviewFormInput) => void;
-  defaultValues?: ReviewFormInput;
-  onHide: () => void;
-  open: boolean;
-}
-
-const ReviewForm: React.FC<ReviewFormProps> = ({ submitText, onSubmit, defaultValues, onHide, open }) => {
-  const {
-    handleSubmit,
-    control,
-    setValue,
-    formState: { errors }
-  } = useForm<ReviewFormInput>({
-    defaultValues: {
-      partId: defaultValues?.partId ?? '',
-      fileIds: defaultValues?.fileIds ?? [],
-      name: defaultValues?.name ?? '',
-      notes: defaultValues?.notes ?? ''
-    },
-    resolver: yupResolver(schema)
-  });
-
-  const {
-    data: manufactuers,
-    isLoading: isLoadingManufactuers,
-    isError: manufacturersIsError,
-    error: manufacturersError
-  } = useGetAllReviews();
-
-  return (
-    <ReviewFormView
-      onSubmit={onSubmit}
-      allParts={parts}
-      handleSubmit={handleSubmit}
-      submitText={submitText}
-      onHide={onHide}
-      control={control}
-      errors={errors}
-      open={open}
-      setValue={setValue}
-    />
-  );
-};
+import { ReviewFormInput } from './ReviewForm';
 
 export interface ReviewFormViewProps {
   submitText: 'Add' | 'Edit';
@@ -74,6 +12,8 @@ export interface ReviewFormViewProps {
   control: Control<ReviewFormInput, any>;
   errors: FieldErrors<ReviewFormInput>;
   open: boolean;
+  setValue: UseFormSetValue<ReviewFormInput>;
+  reset: UseFormReset<ReviewFormInput>;
 }
 
 const ReviewFormView: React.FC<ReviewFormViewProps> = ({
@@ -83,39 +23,33 @@ const ReviewFormView: React.FC<ReviewFormViewProps> = ({
   onHide,
   control,
   errors,
-  allParts,
   open,
-  setValue
+  setValue,
+  reset
 }) => {
   return (
     <NERFormModal
       open={open}
       onHide={onHide}
-      title={submitText + ' Submission'}
+      title={submitText + ' Review'}
       reset={() => {}}
       handleUseFormSubmit={handleSubmit}
       onFormSubmit={onSubmit}
-      formId={submitText + '-submission'}
+      formId={submitText + '-review'}
       showCloseButton
     >
       <FormControl fullWidth>
         <FormLabel>Part</FormLabel>
       </FormControl>
       <FormControl fullWidth sx={{ mb: '10px' }}>
-        <FormLabel>Submission Name</FormLabel>
-        <ReactHookTextField
-          control={control}
-          name="name"
-          errorMessage={errors.name}
-          placeholder="Name..."
-          sx={{ width: 1 }}
-        />
+        <FormLabel>Part</FormLabel> //fix to be dropdown
+        <ReactHookTextField control={control} name="Part" errorMessage={errors.partId} placeholder=" " sx={{ width: 1 }} />
       </FormControl>
       <FormControl fullWidth>
         <FormLabel>Uploader Notes (optional)</FormLabel>
         <ReactHookTextField
           control={control}
-          name="pdmFileName"
+          name="comments"
           errorMessage={errors.notes}
           placeholder="Any additional comments go here..."
           sx={{ width: 1 }}
